Introduce cmd package
