+++ cmd/version.go
@@ -1,0 +1,4 @@
+package cmd
+
+// Version is application version
+const Version string = "0.7.1"
