+++ main.go
@@ -10,35 +10,33 @@ import (
 
 	"github.com/hashicorp/logutils"
 	colorable "github.com/mattn/go-colorable"
+	"github.com/wata727/tflint/cmd"
 )
 
 func main() {
-	cli := &CLI{
-		outStream: colorable.NewColorable(os.Stdout),
-		errStream: colorable.NewColorable(os.Stderr),
-		testMode:  false,
-	}
+	cli := cmd.NewCLI(colorable.NewColorable(os.Stdout), colorable.NewColorable(os.Stderr))
 	filter := &logutils.LevelFilter{
 		Levels:   []logutils.LogLevel{"DEBUG", "INFO", "WARN", "ERROR"},
 		MinLevel: logutils.LogLevel(strings.ToUpper(os.Getenv("TFLINT_LOG"))),
-		Writer:   cli.errStream,
+		Writer:   os.Stderr,
 	}
 	log.SetOutput(filter)
+	log.SetFlags(log.Ltime | log.Lshortfile)
 
 	defer func() {
 		if r := recover(); r != nil {
-			fmt.Fprintf(cli.errStream, "Panic: %v\n", r)
+			fmt.Fprintf(os.Stderr, "Panic: %v\n", r)
 			for depth := 0; ; depth++ {
 				pc, src, line, ok := runtime.Caller(depth)
 				if !ok {
 					break
 				}
-				fmt.Fprintf(cli.errStream, " -> %d: %s: %s(%d)\n", depth, runtime.FuncForPC(pc).Name(), strings.Replace(src, path.Dir(src), "", 1), line)
+				fmt.Fprintf(os.Stderr, " -> %d: %s: %s(%d)\n", depth, runtime.FuncForPC(pc).Name(), strings.Replace(src, path.Dir(src), "", 1), line)
 			}
-			fmt.Fprintln(cli.errStream, `
+			fmt.Fprintln(os.Stderr, `
 TFLint crashed... :(
 Please attach an output log, describe the situation and version that occurred and post an issue to https://github.com/wata727/tflint/issues`)
-			os.Exit(ExitCodeError)
+			os.Exit(cmd.ExitCodeError)
 		}
 	}()
