+++ tflint/runner_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/hashicorp/terraform/configs"
 	"github.com/hashicorp/terraform/terraform"
 	"github.com/k0kubun/pp"
-	"github.com/wata727/tflint/config"
 	"github.com/zclconf/go-cty/cty"
 )
 
@@ -152,7 +151,7 @@ resource "null_resource" "test" {
 			t.Fatal(err)
 		}
 
-		runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+		runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 
 		var ret string
 		err = runner.EvaluateExpr(attribute.Expr, &ret)
@@ -219,7 +218,7 @@ resource "null_resource" "test" {
 			t.Fatal(err)
 		}
 
-		runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+		runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 
 		var ret int
 		err = runner.EvaluateExpr(attribute.Expr, &ret)
@@ -270,7 +269,7 @@ resource "null_resource" "test" {
 			t.Fatal(err)
 		}
 
-		runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+		runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 
 		ret := []string{}
 		err = runner.EvaluateExpr(attribute.Expr, &ret)
@@ -321,7 +320,7 @@ resource "null_resource" "test" {
 			t.Fatal(err)
 		}
 
-		runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+		runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 
 		ret := []int{}
 		err = runner.EvaluateExpr(attribute.Expr, &ret)
@@ -380,7 +379,7 @@ resource "null_resource" "test" {
 			t.Fatal(err)
 		}
 
-		runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+		runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 
 		ret := mapObject{}
 		err = runner.EvaluateExpr(attribute.Expr, &ret)
@@ -477,7 +476,7 @@ resource "null_resource" "test" {
 			t.Fatal(err)
 		}
 
-		runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+		runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 
 		var ret string
 		err = runner.EvaluateExpr(attribute.Expr, &ret)
@@ -730,7 +729,7 @@ resource "null_resource" "test" {
 			t.Fatal(err)
 		}
 
-		runner := NewRunner(config.Init(), cfg, tc.InputValues...)
+		runner := NewRunner(EmptyConfig(), cfg, tc.InputValues...)
 
 		var ret string
 		err = runner.EvaluateExpr(attribute.Expr, &ret)
@@ -772,7 +771,7 @@ func Test_NewModuleRunners_noModules(t *testing.T) {
 		t.Fatalf("Unexpected error occurred: %s", err)
 	}
 
-	runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+	runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 	runners, err := NewModuleRunners(runner)
 	if err != nil {
 		t.Fatalf("Unexpected error occurred: %s", err)
@@ -804,7 +803,7 @@ func Test_NewModuleRunners_nestedModules(t *testing.T) {
 		t.Fatalf("Unexpected error occurred: %s", err)
 	}
 
-	runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+	runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 	runners, err := NewModuleRunners(runner)
 	if err != nil {
 		t.Fatalf("Unexpected error occurred: %s", err)
@@ -972,7 +971,7 @@ func Test_NewModuleRunners_withInvalidExpression(t *testing.T) {
 		t.Fatalf("Unexpected error occurred: %s", err)
 	}
 
-	runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+	runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 	_, err = NewModuleRunners(runner)
 
 	errText := "Failed to eval an expression in module.tf:4; Invalid \"terraform\" attribute: The terraform.env attribute was deprecated in v0.10 and removed in v0.12. The \"state environment\" concept was rename to \"workspace\" in v0.12, and so the workspace name can now be accessed using the terraform.workspace attribute."
@@ -1018,7 +1017,7 @@ func Test_NewModuleRunners_withNotAllowedAttributes(t *testing.T) {
 		t.Fatalf("Unexpected error occurred: %s", err)
 	}
 
-	runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+	runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 	_, err = NewModuleRunners(runner)
 
 	errText := "Attribute of module not allowed was found in module.tf:1; <nil>: Unexpected invalid block; Blocks are not allowed here."
@@ -1080,7 +1079,7 @@ resource "aws_route" "r" {
 		t.Fatal(err)
 	}
 
-	runner := NewRunner(config.Init(), cfg, map[string]*terraform.InputValue{})
+	runner := NewRunner(EmptyConfig(), cfg, map[string]*terraform.InputValue{})
 	resources := runner.LookupResourcesByType("aws_instance")
 
 	if len(resources) != 1 {
