--- version.go
@@ -1,7 +1,0 @@
-package main
-
-// Name is application name
-const Name string = "TFLint"
-
-// Version is application version
-const Version string = "0.7.1"
